+++ packages/zod-validator/src/index.ts
@@ -1,4 +1,4 @@
-import type { Context, MiddlewareHandler, Env, ValidationTargets, TypedResponse } from 'hono'
+import type { Context, MiddlewareHandler, Env, ValidationTargets, TypedResponse, Input } from 'hono'
 import { validator } from 'hono/validator'
 import type { z, ZodSchema, ZodError } from 'zod'
 
@@ -14,20 +14,33 @@ export const zValidator = <
   Target extends keyof ValidationTargets,
   E extends Env,
   P extends string,
-  I = z.input<T>,
-  O = z.output<T>,
-  V extends {
-    in: HasUndefined<I> extends true ? { [K in Target]?: I } : { [K in Target]: I }
-    out: { [K in Target]: O }
-  } = {
-    in: HasUndefined<I> extends true ? { [K in Target]?: I } : { [K in Target]: I }
-    out: { [K in Target]: O }
-  }
+  In = z.input<T>,
+  Out = z.output<T>,
+  I extends Input = {
+    in: HasUndefined<In> extends true
+      ? {
+          [K in Target]?: K extends 'json'
+            ? In
+            : HasUndefined<keyof ValidationTargets[K]> extends true
+            ? { [K2 in keyof In]?: ValidationTargets[K][K2] }
+            : { [K2 in keyof In]: ValidationTargets[K][K2] }
+        }
+      : {
+          [K in Target]: K extends 'json'
+            ? In
+            : HasUndefined<keyof ValidationTargets[K]> extends true
+            ? { [K2 in keyof In]?: ValidationTargets[K][K2] }
+            : { [K2 in keyof In]: ValidationTargets[K][K2] }
+        }
+    out: { [K in Target]: Out }
+  },
+  V extends I = I
 >(
   target: Target,
   schema: T,
   hook?: Hook<z.infer<T>, E, P>
 ): MiddlewareHandler<E, P, V> =>
+  // @ts-expect-error not typed well
   validator(target, async (value, c) => {
     const result = await schema.safeParseAsync(value)