+++ packages/react-renderer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @hono/react-renderer
 
+## 0.1.1
+
+### Patch Changes
+
+- [#419](https://github.com/honojs/middleware/pull/419) [`6eaf2a14cd3bdc1456903e3f9b34ca3964063f79`](https://github.com/honojs/middleware/commit/6eaf2a14cd3bdc1456903e3f9b34ca3964063f79) Thanks [@yoshikouki](https://github.com/yoshikouki)! - fix: overwriting headers when stream is enabled
+
+## 0.1.0
+
+### Minor Changes
+
+- [#410](https://github.com/honojs/middleware/pull/410) [`bda918e286d5e656807b115a9f0aabadc17faa04`](https://github.com/honojs/middleware/commit/bda918e286d5e656807b115a9f0aabadc17faa04) Thanks [@yoshikouki](https://github.com/yoshikouki)! - Support renderToReadableStream options.
+
 ## 0.0.2
 
 ### Patch Changes